move examples to separate files
