--- examples/azure_service_bus/command.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from datetime import timedelta
-
-from diator.requests import Request
-
-
-@dataclass(frozen=True, kw_only=True)
-class CleanUnactiveUsersCommand(Request):
-    eta: timedelta
