--- examples/azure_service_bus/event.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-from diator.events import NotificationEvent
-
-
-@dataclass(frozen=True, kw_only=True)
-class UnactiveUsersCleaned(NotificationEvent):
-    ids: list
