--- examples/di_lib_as_di_container/user_joined_event_handler.py
@@ -1,8 +1,0 @@
-from diator.events import EventHandler
-
-from .user_joined_domain_event import UserJoinedDomainEvent
-
-
-class UserJoinedEventHandler(EventHandler[UserJoinedDomainEvent]):
-    async def handle(self, event: UserJoinedDomainEvent) -> None:
-        print("READY", event)
