--- examples/di_lib_as_di_container/user_joined_notification_event.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass, field
-
-from diator.events import NotificationEvent
-
-
-@dataclass(frozen=True, kw_only=True)
-class UserJoinedNotificationEvent(NotificationEvent):
-    user_id: int = field()
