--- examples/rodi_as_di_container/join_meeting_room_command.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass, field
-
-from diator.requests import Request
-
-
-@dataclass(frozen=True, kw_only=True)
-class JoinMeetingRoomCommand(Request):
-    user_id: int = field()
