--- examples/rodi_as_di_container/join_meeting_room_command_handler.py
@@ -1,18 +1,0 @@
-from diator.requests import RequestHandler
-
-from .join_meeting_room_command import JoinMeetingRoomCommand
-from .user_joined_domain_event import UserJoinedDomainEvent
-from .user_joined_notification_event import UserJoinedNotificationEvent
-
-
-class JoinMeetingRoomCommandHandler(RequestHandler[JoinMeetingRoomCommand, None]):
-    def __init__(self) -> None:
-        self._events = []
-
-    @property
-    def events(self) -> list:
-        return self._events
-
-    async def handle(self, request: JoinMeetingRoomCommand) -> None:
-        self._events.append(UserJoinedDomainEvent(user_id=request.user_id))
-        self._events.append(UserJoinedNotificationEvent(user_id=123))
