--- examples/rodi_as_di_container/user_joined_domain_event.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass, field
-
-from diator.events import DomainEvent
-
-
-@dataclass(frozen=True, kw_only=True)
-class UserJoinedDomainEvent(DomainEvent):
-    user_id: int = field()
